Moving test problems in their own files
