+++ test/src/FENonlinearProblem_test.cpp
@@ -1,7 +1,10 @@
+#include "gmock/gmock.h"
 #include "GodzillaConfig.h"
+#include "FENonlinearProblem_test.h"
 #include "CallStack.h"
 #include "Factory.h"
-#include "FENonlinearProblem_test.h"
+#include "GTestFENonlinearProblem.h"
+#include "GTest2FieldsFENonlinearProblem.h"
 #include "InputParameters.h"
 #include "InitialCondition.h"
 #include "BoundaryCondition.h"
@@ -11,86 +14,28 @@
 using namespace ::testing;
 using namespace godzilla;
 
-registerObject(GTestFENonlinearProblem);
-registerObject(GTest2FieldsFENonlinearProblem);
+/// Test IC with 2 components
+class GTest2CompIC : public InitialCondition {
+public:
+    GTest2CompIC(const InputParameters & params) : InitialCondition(params) {}
+
+    virtual PetscInt
+    get_num_components() const
+    {
+        return 2;
+    }
+
+    virtual void
+    evaluate(PetscInt dim, PetscReal time, const PetscReal x[], PetscInt Nc, PetscScalar u[])
+    {
+        u[0] = 0.;
+        u[1] = 10.;
+    }
+};
+
+
 registerObject(GTest2CompIC);
 
-static void
-f0_u(PetscInt dim,
-     PetscInt Nf,
-     PetscInt NfAux,
-     const PetscInt uOff[],
-     const PetscInt uOff_x[],
-     const PetscScalar u[],
-     const PetscScalar u_t[],
-     const PetscScalar u_x[],
-     const PetscInt aOff[],
-     const PetscInt aOff_x[],
-     const PetscScalar a[],
-     const PetscScalar a_t[],
-     const PetscScalar a_x[],
-     PetscReal t,
-     const PetscReal x[],
-     PetscInt numConstants,
-     const PetscScalar constants[],
-     PetscScalar f0[])
-{
-    f0[0] = 2.0;
-}
-
-/* gradU[comp*dim+d] = {u_x, u_y} or {u_x, u_y, u_z} */
-static void
-f1_u(PetscInt dim,
-     PetscInt Nf,
-     PetscInt NfAux,
-     const PetscInt uOff[],
-     const PetscInt uOff_x[],
-     const PetscScalar u[],
-     const PetscScalar u_t[],
-     const PetscScalar u_x[],
-     const PetscInt aOff[],
-     const PetscInt aOff_x[],
-     const PetscScalar a[],
-     const PetscScalar a_t[],
-     const PetscScalar a_x[],
-     PetscReal t,
-     const PetscReal x[],
-     PetscInt numConstants,
-     const PetscScalar constants[],
-     PetscScalar f1[])
-{
-    PetscInt d;
-    for (d = 0; d < dim; ++d)
-        f1[d] = u_x[d];
-}
-
-/* < \nabla v, \nabla u + {\nabla u}^T >
-   This just gives \nabla u, give the perdiagonal for the transpose */
-static void
-g3_uu(PetscInt dim,
-      PetscInt Nf,
-      PetscInt NfAux,
-      const PetscInt uOff[],
-      const PetscInt uOff_x[],
-      const PetscScalar u[],
-      const PetscScalar u_t[],
-      const PetscScalar u_x[],
-      const PetscInt aOff[],
-      const PetscInt aOff_x[],
-      const PetscScalar a[],
-      const PetscScalar a_t[],
-      const PetscScalar a_x[],
-      PetscReal t,
-      PetscReal u_tShift,
-      const PetscReal x[],
-      PetscInt numConstants,
-      const PetscScalar constants[],
-      PetscScalar g3[])
-{
-    PetscInt d;
-    for (d = 0; d < dim; ++d)
-        g3[d * dim + d] = 1.0;
-}
 
 TEST_F(FENonlinearProblemTest, fields)
 {
@@ -396,56 +341,4 @@ TEST_F(FENonlinearProblemTest, set_constant_2)
     EXPECT_EQ(cs[0], 5);
     EXPECT_EQ(cs[1], 3);
     EXPECT_EQ(cs[2], 1);
-}
-
-// GTestFENonlinearProblem
-
-GTestFENonlinearProblem::GTestFENonlinearProblem(const InputParameters & params) :
-    FENonlinearProblem(params),
-    iu(0)
-{
-}
-
-GTestFENonlinearProblem::~GTestFENonlinearProblem() {}
-
-void
-GTestFENonlinearProblem::set_up_fields()
-{
-    _F_;
-    PetscInt order = 1;
-    add_fe(this->iu, "u", 1, order);
-}
-
-void
-GTestFENonlinearProblem::set_up_weak_form()
-{
-    set_residual_block(this->iu, f0_u, f1_u);
-    set_jacobian_block(this->iu, this->iu, NULL, NULL, NULL, g3_uu);
-}
-
-PetscErrorCode
-GTestFENonlinearProblem::compute_residual_callback(Vec x, Vec f)
-{
-    return FENonlinearProblem::compute_residual_callback(x, f);
-}
-
-PetscErrorCode
-GTestFENonlinearProblem::compute_jacobian_callback(Vec x, Mat J, Mat Jp)
-{
-    return FENonlinearProblem::compute_jacobian_callback(x, J, Jp);
-}
-
-//
-
-GTest2FieldsFENonlinearProblem::GTest2FieldsFENonlinearProblem(const InputParameters & params) :
-    GTestFENonlinearProblem(params),
-    iv(1)
-{
-}
-
-void
-GTest2FieldsFENonlinearProblem::set_up_fields()
-{
-    GTestFENonlinearProblem::set_up_fields();
-    add_fe(this->iv, "v", 1, 1);
 }
